+++ nexus/nxclient/examples/smooth_video.c
@@ -1,7 +1,7 @@
 /***************************************************************************
- *     (c)2015 Broadcom Corporation
+ * Copyright (C) 2017 Broadcom.  The term "Broadcom" refers to Broadcom Limited and/or its subsidiaries.
  *
- * This program is the proprietary software of Broadcom Corporation and/or its licensors,
+ * This program is the proprietary software of Broadcom and/or its licensors,
  * and may only be used, duplicated, modified or distributed pursuant to the terms and
  * conditions of a separate, written license agreement executed between you and Broadcom
  * (an "Authorized License").  Except as set forth in an Authorized License, Broadcom grants
@@ -34,7 +34,6 @@
  * ACTUALLY PAID FOR THE SOFTWARE ITSELF OR U.S. $1, WHICHEVER IS GREATER. THESE
  * LIMITATIONS SHALL APPLY NOTWITHSTANDING ANY FAILURE OF ESSENTIAL PURPOSE OF
  * ANY LIMITED REMEDY.
- *
  **************************************************************************/
 #include "nxclient.h"
 #include "nexus_playback.h"
@@ -300,7 +299,7 @@ int main(int argc, char **argv)
             if (push) {
                 if (total > 1) {
                     NEXUS_SurfaceHandle recycle[MAX_SURFACES];
-                    unsigned num;
+                    size_t num;
 
                     rc = NEXUS_SurfaceClient_RecycleSurface(surfaceClient, recycle, MAX_SURFACES, &num);
                     BDBG_ASSERT(!rc);
